Reindex files.
